requirements.txt
APScheduler==3.10.4
beautifulsoup4==4.12.3
boto3==1.34.66
botocore==1.34.66
Flask==3.0.2
Flask_APScheduler==1.13.1
langchain==0.1.13
langchain_community==0.0.29
langchain_core==0.1.35
langchain_groq==0.0.1
langchain_mongodb==0.1.3
langchain_openai==0.1.1
pymongo==4.6.2
python-dotenv==1.0.1
Requests==2.31.0
streamlit==1.31.1


